Update mod to .rmod format and add icon+banner.
