+++ README.md
@@ -1,11 +1,26 @@
-# SortInventory - Sort your inventory!
+![banner image](./SortInventory/banner.png)
 
-## Description
+# SortInventory - Sort your inventory!
+*([view on RaftModding](https://www.raftmodding.com/mods/sortinventory))*
+
 This is a mod for Raft, a Unity-based game on Steam.  Open your inventory and press 'Z' to have your your inventory (minus hotbar) and any chests you have open to be automatically compressed sorted.
 
 ## Instructions
-1. Get the Raft modloader if you have not already: https://www.raftmodding.com/
-2. Grab the latest release from the releases tab.  Download the DLL and put it in your mods folder.
+1. Get the Raft Mod Loader if you have not already: [https://www.raftmodding.com/download](https://www.raftmodding.com/download)
+2. Click the green `Install mod` button on the right side of this mod's page. This will open the mod installer on your computer.
+
+## Demo
+Before sorting:
+
+![messy inventory](https://github.com/mcglonelevi/RaftModSortInventory/raw/master/messy.PNG)
+
+After sorting:
+![clean inventory](https://github.com/mcglonelevi/RaftModSortInventory/raw/master/clean.PNG)
 
 ## Contributing
-If you discover a bug or want to make an enhancement, please create an issue and/or PR.
+If you discover a bug or want to make an enhancement, please create an issue and/or PR in the [GitHub repo](https://github.com/mcglonelevi/RaftModSortInventory).
+
+## Credits
+- idea and original code by [wazupwiop](https://github.com/mcglonelevi), the original source files are available [on GitHub](https://github.com/mcglonelevi/RaftModSortInventory)
+- new implementation by [traxam](https://trax.am/)
+- icon image based on [an icon image](https://www.flaticon.com/free-icon/sort_178417) by [Freepik](https://www.freepik.com/), licensed [Creative Commons BY 3.0](http://creativecommons.org/licenses/by/3.0/)
