+++ SortInventory/SortInventory.csproj
@@ -34,28 +34,40 @@
     <AppDesigner Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="SortInventory.cs" />
+    <Compile Include="SortInventoryMod.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
-    <Reference Include="Assembly-CSharp">
+    <Reference Include="Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null">
       <HintPath>E:\SteamLibrary\steamapps\common\Raft\Raft_Data/Managed/Assembly-CSharp.dll</HintPath>
+      <HintPath Condition="Exists('C:\Program Files (x86)\Steam\steamapps\common\Raft')">C:\Program Files (x86)\Steam\steamapps\common\Raft\Raft_Data\Managed\Assembly-CSharp.dll</HintPath>
+    </Reference>
+    <Reference Include="AssemblyLoader, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null">
+      <HintPath Condition="Exists('..\..\..\..\AppData\Roaming\RaftModLoader')">..\..\..\..\AppData\Roaming\RaftModLoader\binaries\HMLCoreLibrary.dll</HintPath>
+    </Reference>
+    <Reference Include="coremod, Version=0.0.0.1, Culture=neutral, PublicKeyToken=null">
+      <HintPath>E:\SteamLibrary\steamapps\common\Raft\RaftModLoader/coremod.dll</HintPath>
+      <HintPath Condition="Exists('..\..\..\..\AppData\Roaming\RaftModLoader')">..\..\..\..\AppData\Roaming\RaftModLoader\binaries\coremod.dll</HintPath>
     </Reference>
     <Reference Include="UnityEngine.CoreModule">
       <HintPath>E:\SteamLibrary\steamapps\common\Raft\Raft_Data/Managed/UnityEngine.CoreModule.dll</HintPath>
+      <HintPath Condition="Exists('C:\Program Files (x86)\Steam\steamapps\common\Raft')">C:\Program Files (x86)\Steam\steamapps\common\Raft\Raft_Data\Managed\UnityEngine.CoreModule.dll</HintPath>
     </Reference>
     <Reference Include="Assembly-CSharp-firstpass">
       <HintPath>E:\SteamLibrary\steamapps\common\Raft\Raft_Data/Managed/Assembly-CSharp.dllFIRST</HintPath>
+      <HintPath Condition="Exists('C:\Program Files (x86)\Steam\steamapps\common\Raft')">C:\Program Files (x86)\Steam\steamapps\common\Raft\Raft_Data\Managed\Assembly-CSharp-firstpass.dll</HintPath>
     </Reference>
     <Reference Include="UnityEngine">
       <HintPath>E:\SteamLibrary\steamapps\common\Raft\Raft_Data/Managed/UnityEngine.dll</HintPath>
-    </Reference>
-    <Reference Include="coremod">
-      <HintPath>E:\SteamLibrary\steamapps\common\Raft\RaftModLoader/coremod.dll</HintPath>
+      <HintPath Condition="Exists('C:\Program Files (x86)\Steam\steamapps\common\Raft')">C:\Program Files (x86)\Steam\steamapps\common\Raft\Raft_Data\Managed\UnityEngine.dll</HintPath>
     </Reference>
     <Reference Include="UnityEngine.UI">
       <HintPath>E:\SteamLibrary\steamapps\common\Raft\Raft_Data/Managed/UnityEngine.dll_UI</HintPath>
+      <HintPath Condition="Exists('C:\Program Files (x86)\Steam\steamapps\common\Raft')">C:\Program Files (x86)\Steam\steamapps\common\Raft\Raft_Data\Managed\UnityEngine.UI.dll</HintPath>
     </Reference>
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="modinfo.json" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
