+++ SortInventory/modinfo.json
@@ -1,0 +1,13 @@
+{
+  "name": "SortInventory",
+  "author": "wazupwiop and traxam",
+  "description": "A mod that will use a hotkey to sort inventory.",
+  "version": "1.1.1",
+  "license": "MIT",
+  "icon": "icon.png",
+  "banner": "banner.png",
+  "gameVersion": "Update 11 (4677160)",
+  "updateUrl": "https://www.raftmodding.com/api/v1/mods/sortinventory/version.txt",
+  "isModPermanent": false,
+  "excludedFiles": []
+}
